Use EEPROM and file function calls
Expand the functionality of usbhubctrl by write, read and erase the data of the
EEPROM through the usb hub. The required EEPROM data can be read from a file or
stdin. Furthermore the read EEPROM data can be written to a file or stdout.

This functionality is based on Cpyress USB Hubs CY7C65620/CY7C65630
specification.

Change-Id: I88e406049c85085d6eb49f6bc9a184eedc1f203d
Signed-off-by: Meike Vocke <meike.vocke@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6667
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>