hub-ctrl: Change error handling clear EEPROM
Change the return handling from erase EEPROM. Handle three types of returns.
First the return value is equal to the value that has to be erases.
Second the return value is different to the value that has to be erase.
At least a negative return value, when a error occurs. In the last case
the error usb error message was printed.

Change-Id: Iaf06952f32fe38102b0d28e0dfbb1e4fdf053dfb
Signed-off-by: Meike Vocke <meike.vocke@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6957
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>