tests: use a true library as libusb mock
Put the dummy_usb sources into a static lib and use that as libusb
replacement when linking the tests. This way the implementation is
decided at the linking stage and sources can stay as they are. Also the
mock lib has to use the libusb signatures, enforcing usage of the
original API instead of inventing more or less compatible replacements
as it has been the case before.

Change-Id: Ifa5987c5da7349ec1c6a11d31c90dcb20bae9832
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6982
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>